Migrate ConcurrentNeighborSet to ConcurrentNeighborMap + CNM.Neighbors
This simplifies the concurrency design (only one layer of AtomicReference instead of two) and reduces the per-node overhead from nested instances
